+++ lib/src/regexp.dart
@@ -101,28 +101,183 @@ String regExpReplaceAllMapped(
 /// [multiLine] If [true] the returned [RegExp] will be multiLine.
 /// [caseSensitive] If [true] the returned [RegExp] will be case-sensitive.
 RegExp regExpDialect(Map<String, String> dialectWords, String pattern,
-    {bool multiLine = false, bool caseSensitive = true}) {
-  for (var i = 0; i < 10; i++) {
-    var words2 = dialectWords.map(
-        (k, v) => MapEntry(k, _regExpDialectImpl(dialectWords, v).pattern));
-    if (isEqualsDeep(dialectWords, words2)) break;
-    dialectWords = words2;
+        {bool multiLine = false, bool caseSensitive = true}) =>
+    RegExpDialect(dialectWords,
+            multiLine: multiLine, caseSensitive: caseSensitive)
+        .getPattern(pattern);
+
+/// Represents a dialect. Compiles it on construction.
+class RegExpDialect {
+  Map<String, String> _dialect;
+
+  final bool multiLine;
+
+  final bool caseSensitive;
+
+  RegExpDialect(Map<String, String> dialect,
+      {bool multiLine = false,
+      bool caseSensitive = true,
+      bool throwCompilationErrors = true})
+      : multiLine = multiLine ?? false,
+        caseSensitive = caseSensitive ?? true {
+    _dialect = _compile(dialect, throwCompilationErrors);
   }
-  return _regExpDialectImpl(dialectWords, pattern);
-}
 
-RegExp _regExpDialectImpl(Map<String, String> words, String pattern,
-    {bool multiLine = false, bool caseSensitive = true}) {
-  var translated = pattern.replaceAllMapped(RegExp(r'(\\\$|\$)(\w+)'), (m) {
-    var mark = m.group(1);
-    var key = m.group(2);
+  RegExpDialect._(this._dialect, this.multiLine, this.caseSensitive);
 
-    if (mark == r'\$') return '$mark$key';
+  /// Returns a copy of this instance with parameters [multiLine] and [caseSensitive].
+  /// If this instance already have the same parameters returns this instance.
+  RegExpDialect withParameters(
+      {bool multiLine = false,
+      bool caseSensitive = true,
+      bool throwCompilationErrors = true}) {
+    multiLine ??= false;
+    caseSensitive ??= true;
 
-    var value = words[key];
-    return value ?? '$mark$key';
-  });
-  return RegExp(translated, multiLine: multiLine, caseSensitive: caseSensitive);
+    if (hasErrors && (throwCompilationErrors ?? true)) {
+      throw StateError(
+          "Can't use a dialect with errors. Words with errors: ${errorWords.join(', ')}");
+    }
+
+    if (this.multiLine == multiLine && this.caseSensitive == caseSensitive) {
+      return this;
+    }
+
+    return RegExpDialect._(_dialect, multiLine, caseSensitive);
+  }
+
+  factory RegExpDialect.from(dynamic dialect,
+      {bool multiLine, bool caseSensitive, bool throwCompilationErrors}) {
+    if (dialect == null) return null;
+
+    if (dialect is RegExpDialect) {
+      return dialect.withParameters(
+          multiLine: multiLine,
+          caseSensitive: caseSensitive,
+          throwCompilationErrors: throwCompilationErrors);
+    }
+
+    if (dialect is Map) {
+      var map = asMapOfString(dialect);
+      return RegExpDialect(map,
+          multiLine: multiLine,
+          caseSensitive: caseSensitive,
+          throwCompilationErrors: throwCompilationErrors);
+    }
+
+    return null;
+  }
+
+  Map<String, String> _compile(
+      Map<String, String> dialect, bool throwCompilationErrors) {
+    throwCompilationErrors ??= true;
+
+    for (var i = 0; i < 20; i++) {
+      var words2 = dialect.map((k, v) => MapEntry(
+          k, _compileWordPatternAndSaveErrors(dialect, k, v)?.pattern));
+      if (isEqualsDeep(dialect, words2)) break;
+
+      if (hasErrors) {
+        if (throwCompilationErrors) {
+          throw FormatException(
+              "Error compiling dialect words: '${errorWords.join("', '")}'");
+        } else {
+          return dialect;
+        }
+      }
+
+      dialect = words2;
+    }
+    return dialect;
+  }
+
+  final Map<String, List<String>> _errors = {};
+
+  bool get hasErrors => _errors.isNotEmpty;
+
+  List<String> get errorWords => _errors.keys.toList();
+
+  String getWordErrorMessage(String word) => _errors[word]?.first;
+
+  String getWordErrorPattern(String word) => _errors[word]?.last;
+
+  RegExp _compileWordPatternAndSaveErrors(
+      Map<String, String> dialect, String word, String pattern) {
+    try {
+      return _compileWordPattern(dialect, pattern);
+    } catch (e) {
+      _errors[word] = [e.toString(), pattern];
+      return null;
+    }
+  }
+
+  static final _PATTERN_WORD_PLACEHOLDER = RegExp(r'(\\\$|\$)(\w+)');
+
+  RegExp _compileWordPattern(Map<String, String> words, String pattern) {
+    var translated = pattern.replaceAllMapped(_PATTERN_WORD_PLACEHOLDER, (m) {
+      var mark = m.group(1);
+      var key = m.group(2);
+
+      if (mark == r'\$') return '$mark$key';
+
+      var value = words[key];
+      return value ?? '$mark$key';
+    });
+    return RegExp(translated,
+        multiLine: multiLine, caseSensitive: caseSensitive);
+  }
+
+  List<String> get words => _dialect.keys.toList();
+
+  final Map<String, RegExp> _wordsPatterns = {};
+
+  /// Returns [RegExp] pattern for [word]. Will cache [RegExp] instances.
+  RegExp getWordPattern(String word) {
+    if (word == null) return null;
+    var regexp = _wordsPatterns[word];
+    if (regexp != null) return regexp;
+
+    var pattern = _dialect[word];
+    if (pattern == null) return null;
+
+    if (hasErrors) {
+      if (_errors.containsKey(word)) {
+        throw StateError("Can't use word with compilation error! Word: $word");
+      }
+
+      var patternWords = _PATTERN_WORD_PLACEHOLDER
+          .allMatches(pattern)
+          .map((m) => m.group(2))
+          .toList();
+
+      var patternWordsWithError =
+          patternWords.where((w) => _errors.containsKey(w)).toList();
+
+      if (patternWordsWithError.isNotEmpty) {
+        throw StateError(
+            "Can't use word with compilation error! Words: $patternWordsWithError");
+      }
+    }
+
+    regexp =
+        RegExp(pattern, multiLine: multiLine, caseSensitive: caseSensitive);
+
+    _wordsPatterns[word] = regexp;
+
+    return regexp;
+  }
+
+  final Map<String, RegExp> _patterns = {};
+
+  /// Returns a compiled [RegExp] [pattern], using `$words` in dialect.
+  RegExp getPattern(String pattern) {
+    var regexp = _patterns[pattern];
+    if (regexp != null) return regexp;
+
+    regexp = _compileWordPattern(_dialect, pattern);
+    _patterns[pattern] = regexp;
+    return regexp;
+  }
 }
 
 final RegExp STRING_PLACEHOLDER_PATTERN = RegExp(r'{{(/?\w+(?:/\w+)*/?)}}');